Some Major Progress. Almost done with the planned rollout
Changed the text in the challenge node to bounty instead of challenge, I thought it made it more interesting like "bounty and reward". I placed a limit on how many times the player can change the challenge after which the player spends coins to change the challenge. Changed the pause gui to look more like the menu gui, I'm to make the same changes to the gameover gui. After that i noticed that the same challenge is set after the player changes it so I increment the challenge index by one if that issue comes up but with a slight provision that if the challenge is the last in the table, I decrease it by one. But keep in mind that I haven't fully integrated the challenge system into the game environment.