Customization system progress.
I Set the alphas in the exit prompts back to 1.
I Deleted the prompt in the customization menu gui.
I redesigned the customization menu gui.
I redid the menu shell after deciding to allow customization for box nodes only.
There's now a set all button in the shell collections.
The customization for the menu is now fully functional.
I removed the default colour mechanic.
I felt it unnecessary, so I deleted the code.
I changed "upgrades" to "shop" in the menu gui.
The customization system for the shop is fully functional.