Animations finished and some bugs fixed
I've finished implementing the animations in all the collections. I added more functions to the animation module and changed the scope of the variable BUTTON_ANIMATION_DURATION to the module only. I discovered that the challenge num increases if a bounty has been passed and the reward hasn't been  collected, the challenge num sti;; increases. I added a loop in the init function of the menu script to fix this.