The game is ready
There's no  light mode anymore.. The player can return to the colour menu from the game over collection. The GUI is now centred around a dark background.  Tweak the game play particularly the speed of the objects on screen. I fixed some bugs one relating to the analyser not retrieved cloud data and updating the log in button in the menu after a successful login.